refactor: rename connection_plugin.py to plugin.py
