+++ roscope/plugin.py
